Do not bundle with wsgi_intercept any more
This is available as separately maintained package on PyPI.
The newer versions support requests and don't require the
hack for wsgi_intercept that was included in twill.browser.
Unfortunately, the latest versions break test_wrapper_intercept()
that tests certain legacy write() callable behavior. This should
be fixed in wsgi_intercept if it turns out to be a problem.