Add a command for executing of cleanups
The name of the command  add_cleanup is the same as the similar
method in the unittest package. All added cleanup files will be
run unconditionally after the current script has been finished.
The order will be the same as specified, but the cleanup files
that were added last will be run first. Note that the cleanup
files can actually also be directories of twill files.